Cast serial no for %lld to prevent compiler warnings on some platforms.

+++ usr.bin/ssh/auth2-hostbased.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: auth2-hostbased.c,v 1.54 2025/08/06 04:53:04 djm Exp $ */
+/* $OpenBSD: auth2-hostbased.c,v 1.55 2025/08/14 09:26:53 dtucker Exp $ */
 /*
  * Copyright (c) 2000 Markus Friedl.  All rights reserved.
  *
@@ -216,7 +216,8 @@ hostbased_key_allowed(struct ssh *ssh, struct passwd *pw,
 		    options.fingerprint_hash, SSH_FP_DEFAULT)) == NULL)
 			fatal_f("sshkey_fingerprint fail");
 		error("Refusing certificate ID \"%s\" serial=%llu signed by "
-		    "%s CA %s: %s", key->cert->key_id, key->cert->serial,
+		    "%s CA %s: %s", key->cert->key_id,
+		    (unsigned long long)key->cert->serial,
 		    sshkey_type(key->cert->signature_key), fp, reason);
 		auth_debug_add("Refused Certificate ID \"%s\" serial=%llu: %s",
 		    key->cert->key_id, (unsigned long long)key->cert->serial,
